+++ src/main/scala/ai/starlake/extract/BigQueryTableInfo.scala
@@ -121,7 +121,7 @@ object BigQueryTableInfo extends StrictLogging {
 
     val jobResult = job.run()
     jobResult match {
-      case scala.util.Success(SparkJobResult(Some(dfDataset))) =>
+      case scala.util.Success(SparkJobResult(Some(dfDataset), _)) =>
         BigQuerySparkWriter.sinkInAudit(
           dfDataset,
           "dataset_info",